pyproject.toml
[tool.poetry]
name = "simple-repo-ranker"
version = "0.1.0"
description = ""
authors = ["Andrey Abramov <89972043+anabramo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "0.27.2"
pygithub = "1.58.1"
requests = "2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
